moved sidebar to right
