+++ new-frontend/src/index.css
@@ -11,7 +11,7 @@ body {
   margin: 0;
   transform: scale(0.9);
   /* border: 1px solid #ececec; */
-  border-radius: 35px;
+  border-radius: 30px;
   -webkit-app-region: drag;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
