+++ recognition/app.py
@@ -18,7 +18,7 @@ app = Flask(__name__)
 
 
 # Getting openCV ready
-cap = cv2.VideoCapture(1)
+cap = cv2.VideoCapture(0)
 
 # restricting webcam size
 cap.set(cv2.CAP_PROP_FRAME_WIDTH, 640)
